requirements.txt
torch
torchvision
pandas
matplotlib
scikit-learn
pillow
numpy
streamlit
